ui/requirements.txt
dash
dash-bootstrap-components
pandas
azure-storage-blob
pyyaml
jinja2==3.0.3
python-multipart